Map generation - New room and corridor prefabs (#12)
* Map generation first version

* Map generation scene added

* associate script to the scene

* generate corridors

* Map generation algorithm finished

* Added 3D rooms, corridors and colliders to map

* Removed old terrain generation script

* Reduce room height by 1

* Squashed commit of the following:

commit 0499ebcb6af401458670dcf1a19a472b35fb4243
Author: César Moltó Morilla <cmoltomorilla@gmail.com>
Date:   Tue Mar 8 13:46:18 2022 +0100

    Renamed character controller scripts to avoid conflicts with main branch

commit 1acae86535c2bf8dc6f7522ffaeb8487ccccb95d
Merge: 52c01c1 d14fc65
Author: César Moltó Morilla <cmoltomorilla@gmail.com>
Date:   Tue Mar 8 13:27:48 2022 +0100

    Merge branch 'main' into mechanics

commit 52c01c18538bae056a9a8e59112e33a4fcb5ae29
Author: César Moltó Morilla <cmoltomorilla@gmail.com>
Date:   Tue Mar 8 13:12:47 2022 +0100

    Added mechanics scripts to scripts/mechanics

commit d14fc657f7b1cbae1481340456584c7fe0040e85
Author: César Moltó Morilla <cmoltomorilla@gmail.com>
Date:   Fri Mar 4 12:43:12 2022 +0100

    Photon first approach - 3rd try (#6)

    * Photon test

    * Added local and online matchmaking

    * Reworked multiplayer functionality

    Multiplayer functionality has been reworked. Main things to be noted:
     - Player limit (2, 3, 4).
     - Random room name.
     - Random player name.
     - Go back buttons.

    * Scene names changed

    * Renamed scenes folder name


commit 7d5b2120c1d40e04b397cf78391af87f6ab3e9ba
Author: César Moltó Morilla <cmoltomorilla@gmail.com>
Date:   Wed Mar 2 22:27:06 2022 +0100

    Door interaction mechanics almost done

    Just need to polish some details

commit b615c9baee3295e5962c5262cd59042d0a616b91
Author: César Moltó Morilla <cmoltomorilla@gmail.com>
Date:   Fri Feb 25 16:12:41 2022 +0100

    Added door prefab and fixed character movement bug

commit 3837c4026be0185df070abf6c151492187dba45d
Author: César Moltó Morilla <cmoltomorilla@gmail.com>
Date:   Fri Feb 25 13:41:59 2022 +0100

    Added irene mechanic's scripts

commit 5cb17791cc1f5b691cd9fc5337a78d203c824168
Author: César Moltó Morilla <cmoltomorilla@gmail.com>
Date:   Fri Feb 25 13:16:29 2022 +0100

    Map generation (#3)

    * Map generation first version

    * Map generation scene added

    * associate script to the scene

    * generate corridors

    * Map generation algorithm finished

    * Added 3D rooms, corridors and colliders to map

    * Removed old terrain generation script


commit 828b232fbd544075daa47b02418963f9804e1bed
Author: César Moltó Morilla <cmoltomorilla@gmail.com>
Date:   Fri Feb 25 13:08:30 2022 +0100

    Map generation - 3D rooms, corridors and colliders (#2)

    * Map generation first version

    * Map generation scene added

    * associate script to the scene

    * generate corridors

    * Map generation algorithm finished

    * Added 3D rooms, corridors and colliders to map


* Modified room height and character camera angle

* New door and wall materials added

* Removed old tile and room prefabs

* Modified character prefaba camera angle

+++ Assets/Scripts/TerrainGenerator.cs
@@ -11,8 +11,8 @@ public struct TerrainStructure
     public Vector3 position;
     public Quaternion rotation;
 
-    public int width;
-    public int depth;
+    public float width;
+    public float depth;
 }
 
 public enum CellType
@@ -264,8 +264,8 @@ public class TerrainGenerator : MonoBehaviour
 
             if (temp.Length < 3) { return; }
 
-            int.TryParse(element.prefab.name.Split('_')[1], out element.width); // RandomOddNumber(2, maxRoomWidth);
-            int.TryParse(element.prefab.name.Split('_')[2], out element.depth); // RandomOddNumber(2, maxRoomDepth);
+            float.TryParse(element.prefab.name.Split('_')[1], out element.width); // RandomOddNumber(2, maxRoomWidth);
+            float.TryParse(element.prefab.name.Split('_')[2], out element.depth); // RandomOddNumber(2, maxRoomDepth);
         }
     }
